hopeful fix for circular dependency
